visualisa-platform/backend/pyproject.toml
[project]
name = "visualisa-backend"
version = "1.0.0"
description = "Visualisa Platform Backend — AI-powered renovation visualizer for DR real estate"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.41.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "alembic>=1.18.0",
    "aiomysql>=0.3.0",
    "cryptography>=44.0.0",
    "redis>=7.3.0",
    "celery[redis]>=5.6.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.22",
    "stripe>=14.4.0",
    "boto3>=1.42.0",
    "Pillow>=12.1.0",
    "httpx>=0.28.0",
    "scrapling>=0.4.0",
    "python-slugify>=8.0.4",
    "reportlab>=4.4.0",
    "jinja2>=3.1.6",
    "emails>=0.6.0",
    "tenacity>=9.1.0",
    "structlog>=25.5.0",
    "greenlet>=3.3.0",
    "python-dotenv>=1.2.0",
    "pymysql>=1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.22.0",
    "factory-boy>=3.3.3",
    "faker>=40.0.0",
    "ruff>=0.15.0",
    "mypy>=1.19.0",
    "python-dotenv>=1.2.0",
    "pymysql>=1.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
